+++ tt_metal/api/tt-metalium/bfloat16.hpp
@@ -10,7 +10,7 @@
 #include <functional>
 
 #include <tt-metalium/tile.hpp>
-#include <tt-metalium/span.hpp>
+#include <tt_stl/span.hpp>
 
 class bfloat16 {
 private:
