+++ tt_metal/api/tt-metalium/bfloat8.hpp
@@ -8,7 +8,7 @@
 #include <optional>
 
 #include <tt-metalium/tile.hpp>
-#include <tt-metalium/span.hpp>
+#include <tt_stl/span.hpp>
 
 // TODO: empty struct to facilitate Tensor template logic. Reconsider how/why templating is supported in Tensor
 struct bfloat8_b {};
