+++ tt_metal/api/tt-metalium/device_pool.hpp
@@ -15,7 +15,7 @@
 #include "assert.hpp"
 #include "device.hpp"
 #include "dispatch_core_common.hpp"
-#include "span.hpp"
+#include <tt_stl/span.hpp>
 #include "umd/device/types/cluster_descriptor_types.h"
 #include "control_plane.hpp"
