+++ tt_metal/api/tt-metalium/dispatch_core_common.hpp
@@ -7,7 +7,7 @@
 #include "core_descriptor.hpp"
 #include "core_coord.hpp"
 #include "data_types.hpp"
-#include "reflection.hpp"
+#include <tt_stl/reflection.hpp>
 
 #include <umd/device/tt_core_coordinates.h>  // CoreType
