+++ tt_metal/api/tt-metalium/host_api.hpp
@@ -11,7 +11,7 @@
 #include "program_impl.hpp"
 #include "device.hpp"
 #include "sub_device_types.hpp"
-#include "span.hpp"
+#include <tt_stl/span.hpp>
 #include "lightmetal_binary.hpp"
 
 /** @file */
