+++ tt_metal/api/tt-metalium/small_vector.hpp
@@ -6,7 +6,7 @@
 
 #include <boost/container/small_vector.hpp>
 
-#include "reflection.hpp"
+#include <tt_stl/reflection.hpp>
 
 namespace tt::stl {
