+++ tt_metal/api/tt-metalium/sub_device.hpp
@@ -9,7 +9,7 @@
 
 #include "core_coord.hpp"
 #include "hal.hpp"
-#include "span.hpp"
+#include <tt_stl/span.hpp>
 
 namespace tt::tt_metal {
