+++ tt_metal/api/tt-metalium/sub_device_manager.hpp
@@ -13,7 +13,7 @@
 #include "memcpy.hpp"
 #include "sub_device.hpp"
 #include "sub_device_types.hpp"
-#include "span.hpp"
+#include <tt_stl/span.hpp>
 
 namespace tt::tt_metal {
