+++ tt_metal/impl/data_format/bfloat8.cpp
@@ -13,7 +13,7 @@
 #include "assert.hpp"
 #include "blockfloat_common.hpp"
 #include "tt_backend_api_types.hpp"
-#include "span.hpp"
+#include <tt_stl/span.hpp>
 #include "tracy/Tracy.hpp"
 
 std::vector<uint32_t> pack_fp32_vec_as_bfp8_tiles(
