+++ tt_metal/impl/device/device.cpp
@@ -34,7 +34,7 @@
 #include <sub_device_manager_tracker.hpp>
 #include <sub_device_manager.hpp>
 #include <sub_device_types.hpp>
-#include <span.hpp>
+#include <tt_stl/span.hpp>
 #include <types.hpp>
 
 #include "impl/dispatch/topology.hpp"
