+++ tt_metal/impl/sub_device/sub_device.cpp
@@ -10,7 +10,7 @@
 #include <core_coord.hpp>
 #include <sub_device.hpp>
 #include <hal.hpp>
-#include <span.hpp>
+#include <tt_stl/span.hpp>
 
 namespace tt::tt_metal {
