+++ tt_metal/impl/sub_device/sub_device_manager_tracker.cpp
@@ -19,7 +19,7 @@
 #include <sub_device.hpp>
 #include <sub_device_manager.hpp>
 #include <sub_device_types.hpp>
-#include <span.hpp>
+#include <tt_stl/span.hpp>
 
 namespace tt::tt_metal {
