+++ tt_stl/tt_stl/indestructible.hpp
