+++ tt_stl/tt_stl/strong_type.hpp
